HW3/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.4
pycocotools>=2.0.6
matplotlib>=3.7.0
scikit-image>=0.19.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
